Add files via upload
This all test cases. I write to verify all the functionality, features and modules of website (Admin360).
In this file there are test cases for 4 modules: 1. Login 
                                                                         2. Dashboard
                                                                         3. Guest House System
                                                                         4. Remote Office attendance system
In short, test cases help us confirm that the website is user-friendly. They act as a guide for testing, ensuring that nothing is missed, and help us identify and fix any problems before the website is used by real people.