Update README.md
Our goal is to make sure the admin 360 system works perfectly and meets all the requirements. To 
achieve this, we will conduct thorough testing of all key features, including the Login module, 
Dashboard module, Guest House system, and Remote Office Attendance system. The primary goal of 
this testing phase is to validate that these components function as expected, perform optimally, and 
offer a seamless user experience. By doing so, we can ensure that the admin 360 platform meets its 
requirements and serves as a reliable tool for administrators.